requirements.txt
aiogram
grpcio
grpcio-tools
black
sqlalchemy[asyncio]
asyncpg
alembic
yookassa
redis
orjson
pydantic
psycopg2-binary
pyyaml
cryptography
qrcode[pil]